+++ src/main/java/com/aesopwow/subsubclipclop/domain/analysis/repository/cohortdouble/CohortDoubleAnalysisRepository.java
@@ -1,0 +1,14 @@
+package com.aesopwow.subsubclipclop.domain.analysis.repository.cohortdouble;
+
+import com.aesopwow.subsubclipclop.domain.analysis.dto.cohortdouble.CohortDoubleAnalysisInsightResponseDto;
+import com.aesopwow.subsubclipclop.domain.analysis.dto.cohortdouble.CohortDoubleAnalysisRemainHeatmapResponseDto;
+import com.aesopwow.subsubclipclop.domain.analysis.dto.cohortdouble.CohortDoubleAnalysisUserDataRequestDto;
+import com.aesopwow.subsubclipclop.domain.analysis.dto.cohortdouble.CohortDoubleAnalysisUserDataResponseDto;
+import com.aesopwow.subsubclipclop.domain.analysis.dto.cohortdouble.CohortDoubleAnalysisVisualizationResponseDto;
+
+public interface CohortDoubleAnalysisRepository {
+    CohortDoubleAnalysisInsightResponseDto fetchInsight(String clusterType);
+    CohortDoubleAnalysisRemainHeatmapResponseDto fetchRemainHeatmap(String clusterType);
+    CohortDoubleAnalysisVisualizationResponseDto fetchVisualization(String clusterType);
+    CohortDoubleAnalysisUserDataResponseDto fetchUserData(CohortDoubleAnalysisUserDataRequestDto requestDto);
+}
