--- src/main/java/com/aesopwow/subsubclipclop/domain/analysis/repository/cohortsingle/CohortSingleAnalysisVisualizationRepository.java
@@ -1,7 +1,0 @@
-package com.aesopwow.subsubclipclop.domain.analysis.repository.cohortsingle;
-
-import com.aesopwow.subsubclipclop.domain.analysis.dto.cohortsingle.CohortSingleAnalysisVisualizationResponseDto;
-
-public interface CohortSingleAnalysisVisualizationRepository {
-    CohortSingleAnalysisVisualizationResponseDto fetchVisualization(String clusterType);
-}
