+++ src/main/java/com/aesopwow/subsubclipclop/domain/analysis/service/cohortdouble/CohortDoubleAnalysisService.java
@@ -1,0 +1,16 @@
+package com.aesopwow.subsubclipclop.domain.analysis.service.cohortdouble;
+
+import com.aesopwow.subsubclipclop.domain.analysis.dto.cohortdouble.CohortDoubleAnalysisInsightRequestDto;
+import com.aesopwow.subsubclipclop.domain.analysis.dto.cohortdouble.CohortDoubleAnalysisInsightResponseDto;
+import com.aesopwow.subsubclipclop.domain.analysis.dto.cohortdouble.CohortDoubleAnalysisRemainHeatmapResponseDto;
+import com.aesopwow.subsubclipclop.domain.analysis.dto.cohortdouble.CohortDoubleAnalysisUserDataRequestDto;
+import com.aesopwow.subsubclipclop.domain.analysis.dto.cohortdouble.CohortDoubleAnalysisUserDataResponseDto;
+import com.aesopwow.subsubclipclop.domain.analysis.dto.cohortdouble.CohortDoubleAnalysisVisualizationRequestDto;
+import com.aesopwow.subsubclipclop.domain.analysis.dto.cohortdouble.CohortDoubleAnalysisVisualizationResponseDto;
+
+public interface CohortDoubleAnalysisService {
+    CohortDoubleAnalysisInsightResponseDto fetchInsight(CohortDoubleAnalysisInsightRequestDto requestDto);
+    CohortDoubleAnalysisRemainHeatmapResponseDto fetchRemainHeatmap(String clusterType);
+    CohortDoubleAnalysisVisualizationResponseDto fetchVisualization(CohortDoubleAnalysisVisualizationRequestDto requestDto);
+    CohortDoubleAnalysisUserDataResponseDto fetchUserData(CohortDoubleAnalysisUserDataRequestDto requestDto);
+}
