--- src/main/java/com/aesopwow/subsubclipclop/domain/analysis/service/cohortsingle/CohortSingleAnalysisInsightService.java
@@ -1,8 +1,0 @@
-package com.aesopwow.subsubclipclop.domain.analysis.service.cohortsingle;
-
-import com.aesopwow.subsubclipclop.domain.analysis.dto.cohortsingle.CohortSingleAnalysisInsightRequestDto;
-import com.aesopwow.subsubclipclop.domain.analysis.dto.cohortsingle.CohortSingleAnalysisInsightResponseDto;
-
-public interface CohortSingleAnalysisInsightService {
-    CohortSingleAnalysisInsightResponseDto fetchInsight(CohortSingleAnalysisInsightRequestDto requestDto);
-}
