--- src/main/java/com/aesopwow/subsubclipclop/domain/analysis/service/cohortsingle/CohortSingleAnalysisRemainHeatmapService.java
@@ -1,7 +1,0 @@
-package com.aesopwow.subsubclipclop.domain.analysis.service.cohortsingle;
-
-import com.aesopwow.subsubclipclop.domain.analysis.dto.cohortsingle.CohortSingleAnalysisRemainHeatmapResponseDto;
-
-public interface CohortSingleAnalysisRemainHeatmapService {
-    CohortSingleAnalysisRemainHeatmapResponseDto fetchSingleAnalysisChart(String clusterType);
-}
